requirements/local.txt
-r base.txt

Werkzeug==2.2.2
Werkzeug[watchdog]==2.2.2
ipdb==0.13.9  # https://github.com/gotcha/ipdb
watchgod==0.8.2  # https://github.com/samuelcolvin/watchgod
remote-pdb==2.1 # https://github.com/ionelmc/python-remote-pdb

# factory-boy==3.2.1  # https://github.com/FactoryBoy/factory_boy
# django-coverage-plugin==2.0.2  # https://github.com/nedbat/django_coverage_plugin
# pytest-django==4.4.0  # https://github.com/pytest-dev/pytest-django
django-extensions==3.2.1  # https://github.com/django-extensions/django-extensions
pytest==7.2.0  # https://github.com/pytest-dev/pytest
pytest-cov==4.0.0
pytest-sugar==0.9.6
pytest-mock==3.10.0
pytest-django==4.5.2
coverage==6.5.0
django-coverage-plugin==2.0.4
factory-boy==3.2.1  # https://github.com/FactoryBoy/factory_boy
